refactor: simplify build and packaging of badge component
