+++ packages/components/badge/global.d.ts
@@ -1,0 +1,9 @@
+import { BsBadge } from './bs-badge';
+import { BsBadgeLink } from './bs-badge-link';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        "bs-badge": BsBadge
+        "bs-badge-link": BsBadgeLink
+    }
+}
