+++ packages/components/badge/index.js
@@ -1,0 +1,3 @@
+export * from './css'
+export { BsBadge } from "./bs-badge.js";
+export { BsBadgeLink } from "./bs-badge-link.js";
