+++ packages/components/badge/tsconfig.json
@@ -9,12 +9,11 @@
         "declaration": true,
         "noEmit": false,
         "emitDeclarationOnly": true,
-        "declarationDir": "./types",
         "strict": false,
         "noImplicitAny": false,
         "noImplicitReturns": true,
-        "noUnusedLocals": true
+        "noUnusedLocals": true,
+        "esModuleInterop": true
     },
-    "include": ["src/**/*.js"],
-    "exclude": ["test/**/*.js"]
+    "files": ["index.js"]
 }
