--- packages/components/badge/types/bs-badge-link.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBadgeLinkCss: import("lit-element").CSSResult;
