--- packages/components/badge/types/bs-badge-reboot.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBadgeRebootCss: import("lit-element").CSSResult;
