--- packages/components/badge/types/bs-badge.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBadgeCss: import("lit-element").CSSResult;
