--- packages/components/badge/types/index.d.ts
@@ -1,4 +1,0 @@
-export { BsBadge } from "./bs-badge.js";
-export { BsBadgeCss } from "./bs-badge.css.js";
-export { BsBadgeLink } from "./bs-badge-link.js";
-export { BsBadgeLinkCss } from "./bs-badge-link.css.js";
